requirements.txt
feedparser
pyyaml

